AI Art Demo Bare Code
With @NickGalindo
Functionality is complete, now needs explanations.
It may look little to simple to do, but we can give them a hint of what
is going on by going on in the background by giving them the link to the
repo.
Someone needs to add the badge for Run on Colab.